[#22] 이름 변경
