+++ Projects/Shared/FirebaseSPMShared/Interface/Interface.swift
@@ -1,0 +1,1 @@
+// Interface module source code
