+++ Projects/Shared/FirebaseSPMShared/Sources/Sources.swift
@@ -1,0 +1,1 @@
+// Implementation module source code
