--- Projects/Shared/FirebaseShared/Interface/FirebaseServiceProtocol.swift
@@ -1,3 +1,0 @@
-import Foundation
-
-public protocol FirebaseServiceProtocol { }
