--- Projects/Shared/FirebaseShared/Sources/FirebaseService.swift
@@ -1,6 +1,0 @@
-import Foundation
-import FirebaseSharedInterface
-
-public class FirebaseService: FirebaseServiceProtocol {
-    public init() { }
-}
